move lightTheme to reearthTheme
