+++ web/src/services/theme/index.ts
@@ -7,7 +7,7 @@ export { default as Provider } from "./provider";
 export { default as PublishedAppProvider } from "./publishedAppProvider";
 export { usePublishTheme, mask, type SceneThemeOptions } from "./publishTheme";
 export { default as GlobalStyles } from "./reearthTheme/common/globalStyles";
-export { default as lightTheme } from "./lightTheme";
+export { default as lightTheme } from "./reearthTheme/lightTheme";
 export { default as darkTheme } from "./reearthTheme/darkTheme";
 
 export type { PublishTheme } from "./types";
